requirements.txt
python-binance
pandas
telethon
watchdog
python-telegram-bot